ADDED: authorization for items

+++ src/infrastructure/datasources/item.postgres.ds.ts
@@ -4,10 +4,19 @@ import { CreateItemDto, CustomError, ItemDatasource, ItemEntity, UpdateItemDto }
 
 export class ItemPostgresDs implements ItemDatasource{
     
+    private async validateUserAccess(listId: string, userId_toValidate: string): Promise<void> {
+        const list = await prisma.list.findUnique({where: {id:listId}})
+        if(list?.userId === userId_toValidate) return;
+        throw CustomError.forbidden((`User: ${userId_toValidate} unauthorized to access to item from list: ${listId}`));
+    }
     
-    async createItem(createItemDto: CreateItemDto): Promise<ItemEntity> {
+    async createItem(createItemDto: CreateItemDto, userId: string): Promise<ItemEntity> {
+        const id = createItemDto.id
         try{ 
-            await this.validateItemExists(createItemDto.id, false);
+            const item = await prisma.item.findUnique({where: {id}})
+            if (item) throw CustomError.badRequest(`Item with id: ${id} already exists`);
+            await this.validateUserAccess(createItemDto.listId, userId)
+
 
             const prismaItem = await prisma.item.create({data: createItemDto})
             console.log("new item created");
@@ -15,50 +24,53 @@ export class ItemPostgresDs implements ItemDatasource{
         }
         catch (err:any) {
             const errorMessage = 'Error creating item in postgres: ' + err.message
-            // console.error(errorMessage);
-            throw new CustomError (errorMessage, (err instanceof CustomError) ? err.statusCode : 500);
+            throw new CustomError(errorMessage, (err.statusCode || 500));
         }
     }
     
-    async updateItem(updateItemDto: UpdateItemDto): Promise<ItemEntity> {
+    async updateItem(updateItemDto: UpdateItemDto, userId: string): Promise<ItemEntity> {
         const id = updateItemDto.id
         try{ 
             //validate that the item exists before updating
-            await this.validateItemExists(id);
+            const item = await prisma.item.findUnique({where: {id}})
+            if (!item) throw new CustomError(`Item with id: ${id} not found`, 404);
+            await this.validateUserAccess(item.listId, userId)
             
             const updatedItem = await prisma.item.update({
                 where: {id},
                 data: updateItemDto.updatedValues})
             console.log(`Item with ${id} updated`);
-            return (updatedItem);
+            return updatedItem;
         }
         catch (err:any) {
             const errorMessage = `Error updating item with id ${id} in postgres: ${err.message}`
-            // console.error(errorMessage);
-            throw new CustomError(errorMessage, (err instanceof CustomError) ? err.statusCode: 500);
+            throw new CustomError(errorMessage, (err.statusCode || 500));
         }
     }
     
-    async getItemById(id: string): Promise<ItemEntity> {
+    async getItemById(id: string, userId: string): Promise<ItemEntity> {
         try {
             const item = await prisma.item.findUnique({where: {id}})
             if (!item) throw new CustomError(`Item with id: ${id} not found`, 404)
+            this.validateUserAccess(item!.listId, userId );
             return item
         }
         catch(err: any){
-            const messageError = `Error fetching item with id: ${id} from postgres: ${err.message}`
-            // console.error(messageError);
-            throw new CustomError(messageError, (err instanceof CustomError) ? err.statusCode: 500);
+            const errorMessage = `Error fetching item with id: ${id} from postgres: ${err.message}`
+            throw new CustomError(errorMessage, (err.statusCode || 500));
         }
     }
 
-    async findItemsByDescription(description: string): Promise<ItemEntity[]> {
+    async findItemsByDescription(description: string,  userId: string): Promise<ItemEntity[]> {
         throw new Error("Method not implemented.");
     }
     
-    async deleteItem(id: string): Promise<void> {
+    async deleteItem(id: string,  userId: string): Promise<void> {
         
-        await this.validateItemExists(id)
+        const item = await prisma.item.findUnique({where: {id}})
+        if (!item) throw new CustomError(`Item with id: ${id} not found`, 404);
+        await this.validateUserAccess(item.listId, userId)
+
         try{ 
             //const deletedItem = 
             await prisma.item.delete({where: {id}});
@@ -66,27 +78,7 @@ export class ItemPostgresDs implements ItemDatasource{
             // return deletedItem;
         }catch(err:any){
             const errorMessage = `Error in postgres deleting item with id: ${id}: ${err.message}`;
-            // console.error(errorMessage);
-            throw new CustomError(errorMessage, (err instanceof CustomError) ? err.statusCode: 500);
+            throw new CustomError(errorMessage, (err.statusCode || 500));
         }
     }
-
-    private validateItemExists = async (id: string, awaitedValue? : boolean) : Promise<boolean> => {
-        if (awaitedValue === undefined) awaitedValue= true;
-        const item = await prisma.item.findUnique({where: {id}})
-        if (awaitedValue &&! item) throw new CustomError(`Item with id: ${id} not found`, 404);
-        if (awaitedValue == false && item) throw new CustomError(`Item with id: ${id} already exists`, 400)
-        
-        return (item) ?  true : false;
-    }
-    // public getAllItems = async () : Promise<ItemEntity[]> => {
-    //     try{
-    //         const items = await prisma.item.findMany();
-    //         return items
-    //     }catch(err) {
-    //         const messageError ='Error fetching all items from postgres: ' + err
-    //         console.error(messageError);
-    //         throw new Error(messageError);
-    //     }
-    // }
 }
